Control corner case in which the user that is exporting subscriptions doesn't have an OpmlExportJobState instance associated.

+++ lib/opml_exporter.rb
@@ -18,9 +18,7 @@ class OPMLExporter
   def self.enqueue_export_job(user)
     Rails.logger.info "Enqueuing export subscriptions job for user #{user.email} - #{user.name}"
     # Destroy the current export job state for the user. This in turn triggers a deletion of any old OPML file for the user.
-    # This is not strictly necessary (just creating a new job state will delete the old one in the current ActiveRecord version),
-    # but I think it's better if something this important is as explicit as possible.
-    user.opml_export_job_state.destroy
+    user.opml_export_job_state.destroy if user.opml_export_job_state.present?
     user.create_opml_export_job_state state: OpmlExportJobState::RUNNING
     ExportSubscriptionsWorker.perform_async user.id
     return nil
